backend/requirements.txt
apiclient==1.0.4
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.1.0
awsebcli==3.20.10
beautifulsoup4==4.12.2
botocore==1.31.49
cement==2.8.2
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.3
comm==0.2.0
contourpy==1.2.0
coverage==7.3.2
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
Django==5.0.4
django-extensions==3.2.3
django-js-asset==2.1.0
django-mptt==0.15.0
django-pytest==0.2.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.14.0
dnspython==2.6.1
drf-spectacular==0.26.5
executing==2.0.1
factory-boy==3.3.0
Faker==20.1.0
filelock==3.13.1
flake8==6.1.0
fonttools==4.45.0
fredapi==0.5.1
frozendict==2.3.8
html5lib==1.1
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.27.0
ipython==8.17.2
jedi==0.19.1
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mongoengine==0.28.2
mplfinance==0.12.10b0
multitasking==0.0.11
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pathspec==0.10.1
peewee==3.17.0
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pycodestyle==2.11.1
pydotplus==2.0.2
pyflakes==3.1.0
Pygments==2.17.2
pymongo==4.6.3
pyparsing==3.1.1
pypiwin32==223
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-factoryboy==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytrends==4.9.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.13.1
semantic-version==2.8.5
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
stack-data==0.6.3
TA-Lib==0.4.28
termcolor==1.1.0
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.25.0
wcwidth==0.1.9
webencodings==0.5.1
whitenoise==6.6.0
yfinance==0.2.32
